requirements.txt
pyttsx3
SpeechRecognition
pywhatkit
pyjokes
wikipedia
requests
python-dotenv
PyAudio
